레트로핏 오픈소스 설치 SearchMapActivity 지도 클릭시 현재 위치 나타나도록.

+++ app/src/main/java/com/naver/mycnex/viewpageapplication/data/Comment.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Data
 @ToString
 public class Comment {
-    private Integer id;
+    private double id;
     private String comment;
     private Integer board_id;
 }
